Start documenting how we collected operator counts
This starts explaining how we collected the operator counts, including a simple example of an ancestry tree.
We should replace this tree, though, so we have width proportional to the number of genes that are transmitted
to the winner. In a perfect world we’d also want a better coloring scheme, but that may not happen now.

#9